Bugfix: relative humidity ≥99.5% was reported as h100 instead of h00
It was seen that in situations where ambient relative humidities exceeded 99.5%, the APRS weather message was encoded with the wrong humidity format:

```
11/20/2024 7:36:25 PM-08:00:FW2332>APRS,TCPXX*,qAX,CWOP-3:/210336z4737.94N/12235.79W_190/000g000t042P000h100b10096L001FW2332 210436Z 00000KT 10SM FEW043 05/05 A2981 RMK AO2 SLP096 T00560056=
```

In this example, observe how humidity is encoded as `h100` instead of the required `h00` for 100% humidity.  This caused naïve APRS parsers to decode this as 10% relative humidity.