Fixed a bug with FP4 types having exponents below zero
Closes #1.  It was observed that measurements were clamped to zero when their true values were only close to zero.  This was root caused down to the way exponents are calculated in the PakbusBinaryStream.cs class.

For normal situations, the exponent calculation works correctly.  For example, FP4 data 0x48F3FF61 should resolve to 244 using an exponent of 8.  This works correctly, since (0x48F3FF61 & 0x3F000000) >> 24 == 8.

When under conditions close to zero, such as in the case of 0x3FCBC6A7 (which should resolve to 0.4°C) has a calculated exponent of _-63_.  This is because (0x3FCBC6A7 & 0x3F000000) >> 24 == 63 with a negative sign bit.

In normal implementations of IEEE-754, the exponent is calculated in a similar manner, but has the value 127 subtracted from it.  Since we are using an exponent with one less bit, we will instead subtract the value 64 from our exponent.  We must also consider the mask for exponents here, and use 0x7F000000 instead of 0x3F000000.  This gives us the correct values:  `3 - 64 == -1 and 72 - 64 == 8.